usages function does not read global var
for_modules removed

+++ analyze.sh
@@ -40,7 +40,7 @@ main() {
     local modules="$TMPDIR/modules.tab" 
     find_modules "$target_dir" "$work_dir" "$modules" >&3
     packages "$modules" >&3
-    usages >&3
+    usages "$modules" >&3
     dependency_tree "${includes:-*}" >&3
     # mvn dependency:analyze |awk "/Used undeclared/{s++} /Unused declared/{s--} s{print}" 
     echo "mvn deps" >&3
@@ -233,11 +233,13 @@ packages() {
 
 usages() {
     echo "Finding usages"
+    [ -f "$1" ] || error "Illegal argument"
+    local modules="$1"
     # One line per apparent actual package dependency:
     local outfile="$TMPDIR/deps-detailed.tsv"
     echo -n "" > "$outfile"
 
-    for_modules | cut -f 1,4,5,6 \
+    cut -f 1,4,5,6 "$modules" \
     | while IFS=$'\t' read id base src resource ;do
         find "$base" \( -path "$src/*" -or -path "$resource/*" \) -type f \
             -exec fgrep --color=never --binary-files=without-match -H -o -f "$TMPDIR/packages.txt" {} \; \
@@ -297,10 +299,6 @@ dependency_tree() {
         | while IFS=$'\t' read pom base ;do
             (cd "$base" && mvn -B -q dependency:tree -Dincludes="$includes" -DoutputType=dot -DoutputFile="$TMPDIR/mvn.dot" -DappendOutput=true)
         done
-}
-
-for_modules() {
-    awk '$2 != "pom"' "$TMPDIR/modules.tab"
 }
 
 # reads mvn.dot and deps.tsv
